ActiWatch_package_theft/requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.1.78
numpy>=1.24.0
pandas>=2.1.0
PyYAML>=6.0.1
